contributions: refactoring
