+++ generators/cmd/contributions/main.go
@@ -20,19 +20,14 @@
 package main
 
 import (
-	"context"
 	"flag"
 	"fmt"
-	"github.com/shurcooL/githubv4"
 	log "github.com/sirupsen/logrus"
-	"golang.org/x/oauth2"
-	"gopkg.in/yaml.v3"
+	"kubevirt.io/community/pkg/contributions"
 	"os"
-	"strings"
-	"time"
 )
 
-type options struct {
+type ContributionReportOptions struct {
 	org             string
 	repo            string
 	username        string
@@ -40,7 +35,7 @@ type options struct {
 	months          int
 }
 
-func (o options) validate() error {
+func (o ContributionReportOptions) validate() error {
 	if o.username == "" {
 		return fmt.Errorf("username is required")
 	}
@@ -50,8 +45,17 @@ func (o options) validate() error {
 	return nil
 }
 
-func gatherOptions() (options, error) {
-	o := options{}
+func (o ContributionReportOptions) MakeGeneratorOptions() contributions.ContributionReportGeneratorOptions {
+	return contributions.ContributionReportGeneratorOptions{
+		Org:             o.org,
+		Repo:            o.repo,
+		GithubTokenPath: o.githubTokenPath,
+		Months:          o.months,
+	}
+}
+
+func gatherContributionReportOptions() (ContributionReportOptions, error) {
+	o := ContributionReportOptions{}
 	fs := flag.NewFlagSet(os.Args[0], flag.ExitOnError)
 	fs.StringVar(&o.org, "org", "kubevirt", "org name")
 	fs.StringVar(&o.repo, "repo", "", "repo name (leave empty to create an org activity report)")
@@ -68,109 +72,31 @@ func init() {
 }
 
 func main() {
-	opts, err := gatherOptions()
+	contributionReportOptions, err := gatherContributionReportOptions()
 	if err != nil {
 		log.Fatalf("error parsing arguments %v: %v", os.Args[1:], err)
 	}
-	if err = opts.validate(); err != nil {
+	if err = contributionReportOptions.validate(); err != nil {
 		log.Fatalf("error validating arguments: %v", err)
 	}
-
-	token, err := os.ReadFile(opts.githubTokenPath)
-	if err != nil {
-		log.Fatalf("failed to use github token path %s: %v", opts.githubTokenPath, err)
-	}
-	src := oauth2.StaticTokenSource(
-		&oauth2.Token{AccessToken: strings.TrimSpace(string(token))},
+	err = generateReport(
+		[]string{contributionReportOptions.username},
+		contributionReportOptions.MakeGeneratorOptions(),
 	)
-	httpClient := oauth2.NewClient(context.Background(), src)
-	graphqlClient := githubv4.NewClient(httpClient)
-
-	xMonthsAgo := time.Now().AddDate(0, -1*opts.months, 0)
-
-	if opts.repo != "" {
-		id, err := getUserId(graphqlClient, opts.username)
-		if err != nil {
-			log.Fatalf("failed to query: %v", err)
-		}
-
-		activity, err := generateUserActivityReportInRepository(graphqlClient, opts.org, opts.repo, opts.username, id, xMonthsAgo)
-		if err != nil {
-			log.Fatalf("failed to query: %v", err)
-		}
-		tempFile, err := os.CreateTemp("/tmp", fmt.Sprintf("user-activity-%s-%s_%s-*.yaml", opts.username, opts.org, opts.repo))
-		if err != nil {
-			log.Fatal(err)
-		}
-		defer tempFile.Close()
-		encoder := yaml.NewEncoder(tempFile)
-		err = encoder.Encode(&activity)
-		if err != nil {
-			log.Fatal(err)
-		}
-		fmt.Printf(`activity log:
-	user:       %s
-	repository: %s/%s
-	since:      %s
-
-	issues
-		created:   %d
-		commented: %d
-	pull requests:
-		reviewed:  %d
-		created:   %d
-		commented: %d
-	commits:       %d
-
-user activity log: %q
-`, opts.username, opts.org, opts.repo, xMonthsAgo.Format(time.DateTime),
-			activity.IssuesCreated.IssueCount,
-			activity.IssuesCommented.IssueCount,
-			activity.PullRequestsReviewed.IssueCount,
-			activity.PullRequestsCreated.IssueCount,
-			activity.PullRequestsCommented.IssueCount,
-			activity.CommitsByUser.DefaultBranchRef.Target.Fragment.History.TotalCount,
-			tempFile.Name(),
-		)
-	} else {
-		id, err := getOrganizationId(graphqlClient, opts.org)
-		if err != nil {
-			log.Fatalf("failed to query: %v", err)
-		}
-
-		activity, err := generateUserActivityReportInOrganization(graphqlClient, id, opts.username, xMonthsAgo)
-		if err != nil {
-			log.Fatalf("failed to query: %v", err)
-		}
-		tempFile, err := os.CreateTemp("/tmp", fmt.Sprintf("user-activity-%s-%s-*.yaml", opts.username, opts.org))
-		if err != nil {
-			log.Fatal(err)
-		}
-		defer tempFile.Close()
-		encoder := yaml.NewEncoder(tempFile)
-		err = encoder.Encode(&activity)
-		if err != nil {
-			log.Fatal(err)
-		}
-		fmt.Printf(`activity log:
-	user:         %s
-	organization: %s
-	since:        %s
-
-	hasContributions:                    %t
-	totalIssueContributions:             %d
-	totalPullRequestContributions:       %d
-	totalPullRequestReviewContributions: %d
-	totalCommitContributions:            %d
-
-user activity log: %q
-`, opts.username, opts.org, xMonthsAgo.Format(time.DateTime),
-			activity.HasAnyContributions,
-			activity.TotalIssueContributions,
-			activity.TotalPullRequestContributions,
-			activity.TotalPullRequestReviewContributions,
-			activity.TotalCommitContributions,
-			tempFile.Name(),
-		)
+	if err != nil {
+		log.Fatalf("failed to generate report: %v", err)
 	}
+}
+
+func generateReport(userNames []string, opts contributions.ContributionReportGeneratorOptions) error {
+	generator, err := contributions.NewContributionReportGenerator(opts)
+	if err != nil {
+		return fmt.Errorf("failed to create report generator: %v", err)
+	}
+	for _, userName := range userNames {
+		if err := generator.GenerateReport(userName); err != nil {
+			return err
+		}
+	}
+	return nil
 }
